Fix #1729: add new category behaviors, move category admin to misago.admin
