+++ misago/admin/templates/misago/admin/categories/delete.html
