+++ misago/admin/templates/misago/admin/categories/list.html
