+++ misago/categories/admin/__init__.py
@@ -1,7 +1,6 @@
 from django.urls import path
 from django.utils.translation import pgettext_lazy
 
-from .views import categories
 from .views.perms import (
     CategoryPermissions,
     CategoryRolesList,
@@ -14,33 +13,13 @@ from .views.perms import (
 
 class MisagoAdminExtension:
     def register_urlpatterns(self, urlpatterns):
-        # Categories section
-        urlpatterns.namespace("categories/", "categories")
-
         # Nodes
         urlpatterns.patterns(
             "categories",
-            path("", categories.CategoriesList.as_view(), name="index"),
-            path("new/", categories.NewCategory.as_view(), name="new"),
-            path("edit/<int:pk>/", categories.EditCategory.as_view(), name="edit"),
-            path(
-                "permissions/<int:pk>/",
-                categories.CategoryPermissionsView.as_view(),
-                name="permissions",
-            ),
             path(
                 "permissions-deprecated/<int:pk>/",
                 CategoryPermissions.as_view(),
                 name="permissions-deprecated",
-            ),
-            path(
-                "move/down/<int:pk>/",
-                categories.MoveDownCategory.as_view(),
-                name="down",
-            ),
-            path("move/up/<int:pk>/", categories.MoveUpCategory.as_view(), name="up"),
-            path(
-                "delete/<int:pk>/", categories.DeleteCategory.as_view(), name="delete"
             ),
         )
 
@@ -65,13 +44,6 @@ class MisagoAdminExtension:
         )
 
     def register_navigation_nodes(self, site):
-        site.add_node(
-            name=pgettext_lazy("admin node", "Categories"),
-            icon="fas fa-sitemap",
-            after="ranks:index",
-            namespace="categories",
-        )
-
         site.add_node(
             name=pgettext_lazy("admin node", "Category permissions"),
             parent="permissions",
